Removed extra output from unit tests when they succeed
Unit tests should not print extra details to stdout, especially if they don't fail.
In our case, compareStrings function was implemented correctly - if it returns true, it doesn't output anything.
But in unit test we call it twice, and one of these calls always fails.
So we needed to implement a workaround to output to a separate stream, and print to stdout only if test fails.