Call std::cout.flush() to ensure that contents are shown
Without either std::endl or flush, streams are not guaranteed to send buffer contents to underlying stream. This means that if something happens later in code - we will miss important output. So proper way of outputting is to either use std::endl at the end of important blocks of output, or call flush when newline is not needed.

Usual way to demo the issues is to insert endless loop after output, but for some reason it doesn't work here - maybe because output is large enough to clear buffer anyway.